making it so that when creating a new user in the DB, a user role will be assigned in the user_role table

+++ src/main/java/ons/group8/domain/User.java
@@ -13,7 +13,7 @@ import javax.persistence.*;
 public class User {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    Long id;
+    Integer id;
     private String email;
     private String firstName;
     private String lastName;
